requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-Login==0.6.3
WTForms==3.1.2
email-validator>=2.0.0
python-dotenv==1.0.1
gunicorn==23.0.0
werkzeug==3.1.3
pytest==8.3.3
pytest-flask==1.3.0
pytest-cov==5.0.0
openpyxl==3.1.5





